@@ -1,9 +1,9 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-main-navigation',
   template: `
-    <nav class="app-main-navigation" role="navigation" title="Burger Menu">
+    <nav class="app-main-navigation" role="navigation" title="Burger Menu" (click)="toggleBurger()">
       <div
         class="app-main-navigation__bar"
         [ngClass]="{
@@ -26,10 +26,27 @@ import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
         }"
       ></div>
     </nav>
+
+    <aside
+      class="app-main-navigation__aside"
+      [ngClass]="{
+        'app-main-navigation__aside--active': isVisible
+      }"
+    ></aside>
   `,
   styleUrls: ['./main-navigation.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class MainNavigationComponent {
+export class MainNavigationComponent implements OnInit {
   @Input() public readonly theme: 'black' | 'white';
+
+  public isVisible: boolean;
+
+  ngOnInit(): void {
+    this.isVisible = false;
+  }
+
+  public toggleBurger(): void {
+    this.isVisible = !this.isVisible;
+  }
 }
